Remove typeorm, just have server, edit script for heroku build
