+++ .gitignore
@@ -43,6 +43,7 @@ build/Release
 
 # Dependency directories
 node_modules/
+dist/
 jspm_packages/
 
 # TypeScript v1 declaration files
