+++ index.ts
@@ -1,6 +1,8 @@
 import * as dotenv from 'dotenv';
-import {createServer} from 'http';
-import {app} from './app';
+import * as bodyParser from 'body-parser';
+import * as express from 'express';
+import {Request, Response} from 'express';
+import * as logger from 'morgan';
 
 // Config to use environment variable
 dotenv.config();
@@ -8,12 +10,18 @@ dotenv.config();
 // Create port
 const port = process.env.PORT || 8000;
 
-app.set('port', port);
+export const app: express.Application = express();
 
-const server = createServer(app);
+app.use(bodyParser.json());
+app.use(bodyParser.urlencoded({extended: true}));
+app.use(logger('dev'));
 
-server.listen(port, () => {
-   console.log('App listening on ' + port);
+app.get('*', (_req: Request, res: Response) =>
+    res.status(200).send({
+        message: 'Welcome to test typescript',
+    }),
+);
+
+app.listen(port, () => {
+    console.log('App listening on ' + port);
 });
-
-
