+++ package.json
@@ -2,13 +2,18 @@
   "name": "restaurant-management-server",
   "version": "0.0.1",
   "description": "Server for All restaurant management application",
-  "main": "index.js",
+  "main": "dist/index.js",
   "scripts": {
-    "start": "ts-node index.ts",
     "test": "echo \"Error: no test specified\" && exit 1",
     "dev": "ts-node-dev --respawn --transpileOnly index.ts",
     "prod": "tsc && node index.js",
-    "lint": "tslint --project ."
+    "lint": "tslint --project .",
+    "build-ts": "tsc",
+    "postinstall": "npm run build-ts",
+    "start": "npm run serve",
+    "serve": "node dist/index.js",
+    "watch-node": "nodemon dist/index.js",
+    "watch-ts": "tsc -w"
   },
   "author": "Hieren Lee",
   "license": "ISC",
@@ -18,8 +23,7 @@
     "express": "^4.16.4",
     "morgan": "^1.9.1",
     "pg": "^7.10.0",
-    "tslint-config-airbnb": "^5.11.1",
-    "typeorm": "^0.2.16"
+    "ts-node": "^8.1.0"
   },
   "devDependencies": {
     "@types/body-parser": "^1.17.0",
@@ -28,7 +32,6 @@
     "@types/morgan": "^1.7.35",
     "@types/node": "^11.13.5",
     "@types/pg": "^7.4.14",
-    "ts-node": "^8.1.0",
     "ts-node-dev": "^1.0.0-pre.32",
     "typescript": "^3.4.4"
   }
