+++ tslint.json
@@ -1,8 +1,7 @@
 {
     "defaultSeverity": "error",
     "extends": [
-        "tslint:recommended",
-        "tslint-config-airbnb"
+        "tslint:recommended"
     ],
     "jsRules": {},
     "rules": {
